abstract all linode detail header logic into seperate component
